requirements.txt
curl_cffi
pytest-asyncio
websocket-client
PyYAML
